TST: test_spectrum.py antithrust
