+++ synphot/tests/test_spectrum_source.py
@@ -1,10 +1,8 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
-"""Test spectrum.py module and related functionalities."""
+"""Test spectrum.py module and related functionalities for source spectrum."""
 
 # STDLIB
 import os
-import shutil
-import tempfile
 import warnings
 
 # THIRD-PARTY
@@ -26,11 +24,11 @@ from astropy.utils.exceptions import AstropyUserWarning
 # LOCAL
 from .test_units import _area, _wave, _flux_jy, _flux_photlam, _flux_vegamag
 from .. import exceptions, units
-from ..compat import ASTROPY_LT_4_0, HAS_SPECUTILS  # noqa
+from ..compat import ASTROPY_LT_4_0
+from ..compat import HAS_SPECUTILS  # noqa
 from ..models import (
     BlackBodyNorm1D, Box1D, ConstFlux1D, Empirical1D, Gaussian1D,
-    GaussianAbsorption1D, GaussianFlux1D, Lorentz1D, RickerWavelet1D,
-    PowerLawFlux1D, get_waveset)
+    GaussianFlux1D, Lorentz1D, RickerWavelet1D, PowerLawFlux1D)
 from ..observation import Observation
 from ..spectrum import SourceSpectrum, SpectralElement
 
@@ -93,32 +91,6 @@ def test_flux_conversion_vega(in_q, out_u, ans):
     i = 0
     result = units.convert_flux(_wave[i], in_q[i], out_u, vegaspec=_vspec)
     assert_quantity_allclose(result, ans[i], rtol=1e-2)
-
-
-@pytest.mark.remote_data
-@pytest.mark.skipif('not HAS_SCIPY')
-@pytest.mark.parametrize(
-    'filtername',
-    ['bessel_j', 'bessel_h', 'bessel_k', 'cousins_r', 'cousins_i',
-     'johnson_u', 'johnson_b', 'johnson_v', 'johnson_r', 'johnson_i',
-     'johnson_j', 'johnson_k'])
-def test_filter(filtername):
-    """Test loading pre-defined bandpass.
-
-    .. note::
-
-        Filter data quality is not checked as it depends on the file.
-
-    """
-    bp = SpectralElement.from_filter(filtername)
-    assert isinstance(bp.model, Empirical1D)
-    assert filtername in bp.meta['expr']
-
-
-def test_filter_exception():
-    """Test SpectralElement from_filter() exception."""
-    with pytest.raises(exceptions.SynphotError):
-        SpectralElement.from_filter('foo')
 
 
 @pytest.mark.skipif('not HAS_SCIPY')
@@ -207,169 +179,6 @@ class TestEmpiricalSourceFromFile:
         assert_quantity_allclose(
             y,
             [0, 3.9135e-14, 4.0209e-14, 3.9169e-14, 0] * units.FLAM, rtol=1e-6)
-
-
-@pytest.mark.skipif('not HAS_SCIPY')
-class TestEmpiricalBandpassFromFile:
-    """This is the most common model used in ASTROLIB PYSYNPHOT."""
-    def setup_class(self):
-        bandfile = get_pkg_data_filename(
-            os.path.join('data', 'hst_acs_hrc_f555w.fits'))
-        self.bp = SpectralElement.from_file(bandfile)
-
-    def test_invalid_flux_unit(self):
-        with pytest.raises(u.UnitsError):
-            SpectralElement(Empirical1D, points=_wave,
-                            lookup_table=_flux_photlam)
-
-    def test_call(self):
-        w = self.bp.model.points[0][5000:5004]
-        y = self.bp(w)
-        np.testing.assert_allclose(
-            w, [6045.1640625, 6045.83203125, 6046.49951172, 6047.16748047])
-        np.testing.assert_allclose(
-            y.value, [0.0920415, 0.09125588, 0.09047068, 0.08968487])
-
-    def test_integrate(self):
-        # Whole range (same as EQUVW)
-        f = self.bp.integrate()
-        assert_quantity_allclose(f, 272.01081629459344 * u.AA)
-
-        # Given range
-        f = self.bp.integrate(wavelengths=_wave)
-        assert_quantity_allclose(f, 1.2062975715374322 * u.AA, rtol=2.5e-6)
-
-    def test_avgwave(self):
-        """Compare AVGWAVE with old SYNPHOT result."""
-        w = self.bp.avgwave()
-        assert_quantity_allclose(w, 5367.9 * u.AA, rtol=1e-5)
-
-    def test_barlam(self):
-        """Test BARLAM (no old SYNPHOT result available)."""
-        w = self.bp.barlam()
-        assert_quantity_allclose(w, 5331.8945 * u.AA, rtol=1e-5)
-
-    def test_pivot(self):
-        """Compare PIVWV with ASTROLIB PYSYNPHOT result."""
-        w = self.bp.pivot()
-        assert_quantity_allclose(w, 5355.863596422962 * u.AA, rtol=1e-6)
-
-    def test_uresp(self):
-        """Compare URESP with old SYNPHOT result."""
-        uresp = self.bp.unit_response(area=_area)
-        assert_quantity_allclose(uresp, 3.00737e-19 * units.FLAM, rtol=1e-4)
-
-    def test_rmswidth(self):
-        w = self.bp.rmswidth()
-        assert_quantity_allclose(w, 359.55954282883687 * u.AA, rtol=1e-4)
-
-        w = self.bp.rmswidth(threshold=0.01 * u.dimensionless_unscaled)
-        assert_quantity_allclose(w, 357.43298216917754 * u.AA, rtol=1e-4)
-
-        # Invalid threshold must raise exception
-        with pytest.raises(exceptions.SynphotError):
-            self.bp.rmswidth(threshold=0.01 * u.AA)
-        with pytest.raises(exceptions.SynphotError):
-            self.bp.rmswidth(threshold=[0.01, 0.02])
-        with pytest.raises(exceptions.SynphotError):
-            self.bp.rmswidth(threshold='foo')
-
-    def test_fwhm(self):
-        """This also calls PHOTBW."""
-        w = self.bp.fwhm()
-        assert_quantity_allclose(w, 841.09 * u.AA, rtol=2.5e-5)
-
-        w = self.bp.fwhm(threshold=0.01 * u.dimensionless_unscaled)
-        assert_quantity_allclose(w, 836.2879507505378 * u.AA, rtol=2.5e-5)
-
-        # Zero value
-        w = self.bp.fwhm(wavelengths=[2e6, 2.1e6])
-        assert w.value == 0
-
-        # Invalid threshold must raise exception
-        with pytest.raises(exceptions.SynphotError):
-            self.bp.fwhm(threshold=0.01 * u.AA)
-        with pytest.raises(exceptions.SynphotError):
-            self.bp.fwhm(threshold=[0.01, 0.02])
-        with pytest.raises(exceptions.SynphotError):
-            self.bp.fwhm(threshold='foo')
-
-    def test_tlambda(self):
-        f = self.bp.tlambda()
-        assert_quantity_allclose(f, 0.22808, rtol=1e-4)
-
-    def test_tpeak(self):
-        """Compare TPEAK with old SYNPHOT result."""
-        f = self.bp.tpeak()
-        assert_quantity_allclose(f, 0.241445)
-
-    def test_wpeak(self):
-        w = self.bp.wpeak()
-        assert_quantity_allclose(w, 5059.8 * u.AA, rtol=1e-5)
-
-    def test_equivwidth(self):
-        """Compare EQUVW with ASTROLIB PYSYNPHOT result."""
-        w = self.bp.equivwidth()
-        assert_quantity_allclose(w, 272.01081629459344 * u.AA, rtol=1e-6)
-
-    def test_rectw(self):
-        """Compare RECTW with old SYNPHOT result."""
-        w = self.bp.rectwidth()
-        assert_quantity_allclose(w, 1126.588 * u.AA, rtol=1e-5)
-
-    def test_qtlam(self):
-        qtlam = self.bp.efficiency()
-        assert_quantity_allclose(qtlam, 0.050901, rtol=1e-4)
-
-    def test_emflx(self):
-        """Compare EMFLX with old SYNPHOT result."""
-        f = self.bp.emflx(area=_area)
-        assert_quantity_allclose(f, 3.586622e-16 * units.FLAM, rtol=2.5e-5)
-
-
-class TestBoxBandpass:
-    """Test bandpass with Box1D model."""
-    def setup_class(self):
-        self.bp = SpectralElement(Box1D, amplitude=1, x_0=5000, width=100)
-
-    def test_eval(self):
-        # Box: Outside, boundary, inside
-        y = self.bp([4000, 4949.95, 5000])
-        np.testing.assert_array_equal(y.value, [0, 0, 1])
-
-    def test_conversion(self):
-        bp2 = SpectralElement(
-            Box1D, amplitude=1, x_0=500 * u.nm, width=10 * u.nm)
-        y = bp2([4000, 4949.95, 5000])
-        np.testing.assert_array_equal(y.value, [0, 0, 1])
-
-    def test_fwhm(self):
-        # You would think FWHM of a box is the width but
-        # not according to IRAF SYNPHOT.
-        assert_quantity_allclose(self.bp.fwhm(), 67.977 * u.AA,
-                                 rtol=1e-3)  # 0.1%
-
-    def test_taper(self):
-        bp2 = self.bp.taper(np.arange(499, 501.01, 0.01) * u.nm)
-        y = bp2([498.9, 499, 500, 501, 501.1] * u.nm)
-        assert_quantity_allclose(y, [0, 1, 1, 1, 0])
-
-    def test_integrate(self):
-        ans = 100 * u.AA
-        assert_quantity_allclose(self.bp.integrate(), ans)
-        assert_quantity_allclose(
-            self.bp.integrate(integration_type='analytical'), ans)
-
-        with pytest.raises(exceptions.SynphotError,
-                           match='flux_unit cannot be used'):
-            self.bp.integrate(integration_type='analytical', flux_unit='flam')
-
-    def test_multi_n_models(self):
-        """This is not allowed."""
-        with pytest.raises(exceptions.SynphotError):
-            SpectralElement(
-                Box1D, amplitude=[1, 1], x_0=[5000, 6000],
-                width=[100, 1], n_models=2)
 
 
 class TestBlackBodySource:
@@ -525,7 +334,7 @@ class TestPowerLawSource:
                               integration_type='analytical'), ans_flam)
 
 
-class TestBuildModels:
+class TestBuildModelsSource:
     """Test compatiblity with other models not tested above."""
     def test_BrokenPowerLaw1D(self):
         sp = SourceSpectrum(
@@ -553,13 +362,6 @@ class TestBuildModels:
         assert_quantity_allclose(
             y, [1.25770198, 0.54881164, 0.04767718] * units.PHOTLAM)
 
-    def test_GaussianAbsorption1D(self):
-        """This should be unitless, not a source spectrum."""
-        bp = SpectralElement(
-            GaussianAbsorption1D, amplitude=0.8, mean=5500, stddev=50)
-        y = bp([5300, 5500, 5700])
-        assert_quantity_allclose(y, [0.99973163, 0.2, 0.99973163])
-
     def test_LogParabola1D(self):
         sp = SourceSpectrum(
             LogParabola1D, amplitude=1, x_0=6000, alpha=1, beta=4)
@@ -582,95 +384,6 @@ class TestBuildModels:
         sp = SourceSpectrum(PowerLaw1D, amplitude=1, x_0=6000, alpha=4)
         y = sp([5000, 6000, 7000])
         assert_quantity_allclose(y, [2.0736, 1, 0.53977509] * units.PHOTLAM)
-
-
-@pytest.mark.skipif('not HAS_SCIPY')
-class TestCheckOverlap:
-    """Test spectrum overlap check. This method is ever only used
-    in the form of ``bp.check_overlap(sp)``, so that is what is
-    tested here.
-    """
-    def setup_class(self):
-        self.bp = SpectralElement(
-            Empirical1D, points=[2999.9, 3000, 6000, 6000.1],
-            lookup_table=[0, 1, 1, 0])
-
-    def test_full(self):
-        """As long as we don't have to extrapolate or taper
-        source spectrum, it's okay.
-        """
-        sp = SourceSpectrum(
-            Empirical1D, points=[999.9, 1000, 9000, 9000.1],
-            lookup_table=[0, 1, 1, 0])
-        assert self.bp.check_overlap(sp) == 'full'
-
-        sp = SourceSpectrum(
-            Empirical1D, points=[3999.9, 4000, 4500, 4500.1],
-            lookup_table=[0, 1, 1, 0])
-        assert self.bp.check_overlap(sp) == 'full'
-
-    def test_partial_most(self):
-        """99% overlap."""
-        sp = SourceSpectrum(
-            Empirical1D, points=[3005, 3005.1, 6000.1, 6000.2],
-            lookup_table=[0, 1, 1, 0])
-        assert self.bp.check_overlap(sp) == 'partial_most'
-
-    def test_partial_notmost(self):
-        """Extrapolation or taper required."""
-        sp = SourceSpectrum(
-            Empirical1D, points=[3999.9, 4500.1], lookup_table=[1, 1])
-        assert self.bp.check_overlap(sp) == 'partial_notmost'
-
-    def test_none(self):
-        """No overlap at all."""
-        sp = SourceSpectrum(
-            Empirical1D, points=[99.9, 100, 2999.8, 2999.9],
-            lookup_table=[0, 1, 1, 0])
-        assert self.bp.check_overlap(sp) == 'none'
-
-    def test_special_cases(self):
-        """One of them has no waveset defined."""
-        # Other has no waveset
-        sp = SourceSpectrum(Const1D, amplitude=1)
-        assert self.bp.check_overlap(sp) == 'full'
-
-        # Self has no waveset
-        bp = SpectralElement(Const1D, amplitude=1)
-        sp = SourceSpectrum(Box1D, amplitude=1, x_0=5000, width=10)
-        assert bp.check_overlap(sp) == 'partial_notmost'
-
-    def test_exceptions(self):
-        """Invalid input."""
-        with pytest.raises(exceptions.SynphotError):
-            self.bp.check_overlap(1)
-
-
-class TestForceExtrap:
-    """Test forcing extrapolation on a source spectrum."""
-
-    @pytest.mark.skipif('not HAS_SCIPY')
-    @pytest.mark.parametrize('z', [0, 0.03])
-    def test_empirical(self, z):
-        sp = SourceSpectrum(Empirical1D, points=[1000, 2000, 3000, 4000],
-                            lookup_table=[0.5, 0.6, 10.6, 1.5], fill_value=0)
-        sp.z = z
-        w = [900, 4300]
-        assert_quantity_allclose(sp(w), 0 * units.PHOTLAM)  # No extrapolation
-
-        is_forced = sp.force_extrapolation()  # Force extrapolation
-        assert is_forced
-        assert_quantity_allclose(sp(w), [0.5, 1.5] * units.PHOTLAM)
-
-    def test_analytical(self):
-        """Forcing is not possible."""
-        sp = SourceSpectrum(GaussianFlux1D, mean=5500, total_flux=1, fwhm=10)
-        w = [100, 10000]
-        assert_quantity_allclose(sp(w), 0 * units.PHOTLAM)
-
-        is_forced = sp.force_extrapolation()
-        assert not is_forced
-        assert_quantity_allclose(sp(w), 0 * units.PHOTLAM)
 
 
 @pytest.mark.skipif('not HAS_SCIPY')
@@ -811,96 +524,6 @@ class TestNormalize:
             sp.normalize(100 * u.ct, band=self.abox, area=_area)
 
 
-class TestWaveset:
-    """Tests related to spectrum waveset."""
-    def test_none(self):
-        sp = SourceSpectrum(Const1D, amplitude=1)
-        assert sp.waveset is None
-
-    def test_sampleset(self):
-        tf_unit = u.erg / (u.cm * u.cm * u.s)
-        sp = SourceSpectrum(
-            GaussianFlux1D, total_flux=(1 * tf_unit), mean=5000, fwhm=10)
-        np.testing.assert_array_equal(sp.waveset.value, sp.model.sampleset())
-
-    def test_box1d(self):
-        bp = SpectralElement(Box1D, x_0=2000, width=1)
-        w = bp.waveset.value
-        w_true = bp.model.sampleset()
-
-        np.testing.assert_array_equal(w, w_true)
-        np.testing.assert_allclose(
-            w[([0, 1, -2, -1], )], bp.model.sampleset(minimal=True))
-
-        # Make sure scale does not change waveset
-        bp2 = bp * 2
-        bp3 = 0.5 * bp
-        np.testing.assert_array_equal(bp2.waveset.value, w_true)
-        np.testing.assert_array_equal(bp3.waveset.value, w_true)
-
-    def test_composite_none(self):
-        bp1 = SpectralElement(Box1D, amplitude=1, x_0=5000, width=10)
-        bp2 = SpectralElement(Const1D, amplitude=2)
-        bp = bp1 * bp2
-        np.testing.assert_array_equal(bp.waveset, bp1.waveset)
-
-    def test_composite(self):
-        totflux = 1 * (u.erg / (u.cm * u.cm * u.s))
-        g1 = SourceSpectrum(
-            GaussianFlux1D, total_flux=totflux, mean=5000, fwhm=10)
-        g2 = SourceSpectrum(
-            GaussianFlux1D, total_flux=totflux, mean=6500, fwhm=100)
-        g3 = SourceSpectrum(
-            GaussianFlux1D, total_flux=totflux, mean=7500, fwhm=5)
-        sp = SpectralElement(Box1D, x_0=1000, width=1) * (g1 + g2 + g3)
-        assert_quantity_allclose(
-            sp.waveset[::100],
-            [999.49, 1000.49, 5019.95906231, 6699.59062307,
-             7509.7672007] * u.AA)
-
-    def test_redshift(self):
-        tf_unit = u.erg / (u.cm * u.cm * u.s)
-        sp = SourceSpectrum(
-            GaussianFlux1D, total_flux=(1 * tf_unit), mean=5000, fwhm=10)
-        sp.z = 1.3
-        m = RedshiftScaleFactor(z=1.3)
-        w_step25_z0 = [4978.76695499, 4989.3834775, 5000, 5010.6165225] * u.AA
-        assert_quantity_allclose(sp.waveset[::25], m(w_step25_z0))
-
-    def test_redshift_none(self):
-        sp = SourceSpectrum(Const1D, amplitude=1, z=1.3)
-        assert sp.waveset is None
-
-    def test_complicated_tree(self):
-        """Throw everything in and insert redshift and scale in the middle."""
-
-        # On one side, we have a composite bandpass.
-        bp1 = SpectralElement(Const1D, amplitude=1.01)
-        bp2 = SpectralElement(
-            Empirical1D, points=[4999, 5000.001, 5030], lookup_table=[0, 1, 0])
-        bp = bp1 * (0.8 * bp2)  # [4999, 5000.001, 5030]
-
-        # On the other side, we have composite spectrum with
-        # scale and redshift.
-        sp1 = SourceSpectrum(
-            Empirical1D, points=[5001, 5011, 5020], lookup_table=[0, 1, 0])
-        sp2 = SourceSpectrum(
-            Empirical1D, points=[5000, 5010, 5020], lookup_table=[0, 1, 0])
-        sp3 = sp2 + (sp1 * 0.5)  # [5000, 5001, 5010, 5011, 5020]
-        sp3.z = 0.01  # [5050, 5051.01, 5060.1, 5061.11, 5070.2]
-        sp = sp1 + sp3  # [5001, 5011, 5020, 5050, 5051.01, 5060.1, 5061.11, 5070.2]  # noqa
-
-        sp_final = sp * bp
-        np.testing.assert_array_equal(
-            sp_final.waveset.value,
-            [4999, 5000.001, 5001, 5011, 5020, 5030, 5050, 5051.01,
-             5060.1, 5061.11, 5070.2])
-
-    def test_exceptions(self):
-        with pytest.raises(exceptions.SynphotError):
-            get_waveset('foo')
-
-
 class TestRedShift:
     """Test redshifted source spectrum.
 
@@ -954,213 +577,8 @@ class TestRedShift:
 
 
 @pytest.mark.skipif('not HAS_SCIPY')
-class TestMathOperators:
-    """Test spectrum math operators."""
-    def setup_class(self):
-        self.sp_1 = SourceSpectrum(
-            Empirical1D, points=[3999.9, 4000.0, 5000.0, 6000.0, 6000.1],
-            lookup_table=[0, 3.5e-14, 4e-14, 4.5e-14, 0] * units.FLAM)
-        self.sp_2 = SourceSpectrum(
-            Empirical1D, points=_wave, lookup_table=_flux_jy,
-            meta={'PHOTLAM': [9.7654e-3, 1.003896e-2, 9.78473e-3]})
-        self.bp_1 = SpectralElement(
-            Empirical1D, points=[399.99, 400.01, 500.0, 590.0, 600.1] * u.nm,
-            lookup_table=[0, 0.1, 0.2, 0.3, 0])
-
-    def test_source_add(self):
-        """Compare with ASTROLIB PYSYNPHOT."""
-        ans = self.sp_1 + self.sp_2
-        assert_quantity_allclose(
-            ans(ans.waveset),
-            [0.00976521, 0.01681283, 0.01970276, 0.01998463, 0.0197387,
-             0.01985257, 0.02337638, 0.00978454] * units.PHOTLAM, rtol=1e-4)
-
-    def test_source_sub(self):
-        """Compare with ASTROLIB PYSYNPHOT."""
-        ans = self.sp_1 - self.sp_2
-        assert_quantity_allclose(
-            ans(ans.waveset),
-            [-9.76520783e-03, -2.71758275e-03, 1.72346256e-04, -9.29051118e-05,
-             1.69629843e-04, 2.83499328e-04, 3.80731187e-03,
-             -9.78453651e-03] * units.PHOTLAM,
-            rtol=1e-4)
-
-    def test_source_addsub_circular(self):
-        """sp = sp + sp - sp"""
-        ans = self.sp_1 + self.sp_1 - self.sp_1
-        assert_quantity_allclose(ans(ans.waveset), self.sp_1(ans.waveset))
-
-    def test_source_addsub_exception(self):
-        with pytest.raises(exceptions.IncompatibleSources):
-            self.sp_1 + self.bp_1
-
-    @pytest.mark.parametrize('x', [2, 2 * u.dimensionless_unscaled])
-    def test_source_mul_div_scalar(self, x):
-        w = self.sp_1.waveset
-
-        ans1 = self.sp_1 * x
-        assert_quantity_allclose(
-            ans1(w),
-            [0, 0.01409552, 0.02013646, 0.02718424, 0] * units.PHOTLAM,
-            rtol=1e-6)
-
-        # rmul does not work with Quantity
-        if not isinstance(x, u.Quantity):
-            ans2 = x * self.sp_1
-            assert_quantity_allclose(ans1(w), ans2(w), rtol=0)
-
-        ans3 = self.sp_1 / x
-        assert_quantity_allclose(
-            ans3(w),
-            [0, 0.00352388, 0.00503411, 0.00679606, 0] * units.PHOTLAM,
-            atol=1e-7 * units.PHOTLAM)
-
-    def test_source_mul_div_spec(self):
-        """Compare mul with ASTROLIB PYSYNPHOT. Also test bp * sp."""
-        ans1 = self.sp_1 * self.bp_1
-        ans2 = self.bp_1 * self.sp_1
-        w = ans1.waveset[:-1]
-        assert_quantity_allclose(
-            ans1(w),
-            [0, 3.52381254e-04, 7.04792712e-04, 2.01360717e-03, 3.97184014e-03,
-             4.03718269e-05, 0] * units.PHOTLAM, rtol=1e-4)
-        assert_quantity_allclose(ans1(w), ans2(w), rtol=0)
-
-        ans3 = self.sp_1 / self.bp_1
-        assert_quantity_allclose(
-            ans3(w),
-            [0, 0.14095528, 0.07048066, 0.05034117, 0.04413243, 4.57601236,
-             0] * units.PHOTLAM)
-
-        ans4 = self.sp_1 / self.sp_1
-        assert_quantity_allclose(
-            ans4([4000, 5000, 6000]), 1 * u.dimensionless_unscaled)
-
-        # Dividing throughput by flux does not make sense.
-        with pytest.raises(exceptions.IncompatibleSources):
-            self.bp_1 / self.sp_1
-
-    def test_source_mul_div_exceptions(self):
-        """Only mul is tested but truediv uses the same validation."""
-        with pytest.raises(exceptions.IncompatibleSources):
-            self.sp_1 * self.sp_2
-
-        with pytest.raises(exceptions.IncompatibleSources):
-            self.sp_1 * [1, 2]
-        with pytest.raises(exceptions.IncompatibleSources):
-            self.sp_1 * (1 - 1j)
-
-        with pytest.raises(exceptions.IncompatibleSources):
-            self.sp_1 * u.Quantity([1, 2])
-        with pytest.raises(exceptions.IncompatibleSources):
-            self.sp_1 * u.Quantity(1 - 1j)
-        with pytest.raises(exceptions.IncompatibleSources):
-            self.sp_1 * (1 * u.AA)
-
-    def test_bandpass_addsub(self):
-        """Not supported."""
-        with pytest.raises(NotImplementedError):
-            self.bp_1 + self.bp_1
-        with pytest.raises(NotImplementedError):
-            self.bp_1 + 2.0
-        with pytest.raises(NotImplementedError):
-            self.bp_1 - self.bp_1
-        with pytest.raises(NotImplementedError):
-            self.bp_1 - 2.0
-
-    @pytest.mark.parametrize('x', [2.0, 2.0 * u.dimensionless_unscaled])
-    def test_bandpass_mul_div_scalar(self, x):
-        w = self.bp_1.waveset
-
-        ans1 = self.bp_1 * x
-        assert_quantity_allclose(ans1(w), [0, 0.2, 0.4, 0.6, 0])
-
-        # rmul does not work with Quantity
-        if not isinstance(x, u.Quantity):
-            ans2 = x * self.bp_1
-            assert_quantity_allclose(ans1(w), ans2(w), rtol=0)
-
-        ans3 = self.bp_1 / x
-        assert_quantity_allclose(ans3(w), [0, 0.05, 0.1, 0.15, 0])
-
-    def test_bandpass_mul_div_bandpass(self):
-        ans1 = self.bp_1 * self.bp_1
-        assert_quantity_allclose(
-            ans1(ans1.waveset), [0, 0.01, 0.04, 0.09, 0])
-
-        w = [4000.1, 5000, 5900]  # Avoid div by zero
-        ans2 = self.bp_1 / self.bp_1
-        assert_quantity_allclose(ans2(w), 1)
-
-    def test_bandpass_mul_div_exceptions(self):
-        """Only mul is tested but truediv uses the same validation."""
-        class DummyObject:
-            pass
-
-        with pytest.raises(exceptions.IncompatibleSources):
-            self.bp_1 * DummyObject()
-
-        with pytest.raises(exceptions.IncompatibleSources):
-            self.bp_1 * u.Quantity([1, 2])
-        with pytest.raises(exceptions.IncompatibleSources):
-            self.bp_1 * u.Quantity(1 - 1j)
-        with pytest.raises(exceptions.IncompatibleSources):
-            self.bp_1 * (1 * u.AA)
-
-        with pytest.raises(exceptions.IncompatibleSources):
-            self.bp_1 * [1, 2]
-        with pytest.raises(exceptions.IncompatibleSources):
-            self.bp_1 * (1 - 1j)
-
-
-@pytest.mark.skipif('not HAS_SCIPY')
-class TestWriteSpec:
-    """Test spectrum to_fits() method."""
-    def setup_class(self):
-        self.outdir = tempfile.mkdtemp()
-        self.sp = SourceSpectrum(
-            Empirical1D, points=_wave, lookup_table=_flux_photlam,
-            meta={'expr': 'Test source'})
-        self.bp = SpectralElement(
-            Empirical1D, points=_wave, lookup_table=np.ones(_wave.shape),
-            meta={'expr': 'Test bandpass'})
-
-    @pytest.mark.parametrize(
-        ('is_sp', 'ext_hdr'),
-        [(True, None),
-         (True, {'foo': 'foo'}),
-         (False, None),
-         (False, {'foo': 'foo'})])
-    def test_write(self, is_sp, ext_hdr):
-        outfile = os.path.join(self.outdir, 'outspec.fits')
-
-        if is_sp:
-            sp1 = self.sp
-        else:
-            sp1 = self.bp
-
-        if ext_hdr is None:
-            sp1.to_fits(outfile, overwrite=True, trim_zero=False,
-                        pad_zero_ends=False)
-        else:
-            sp1.to_fits(outfile, overwrite=True, trim_zero=False,
-                        pad_zero_ends=False, ext_header=ext_hdr)
-
-        # Read it back in and check
-        sp2 = sp1.__class__.from_file(outfile)
-        np.testing.assert_allclose(sp2(sp2.waveset), sp1(sp1.waveset))
-        hdr = fits.getheader(outfile, 1)
-        assert 'expr' in hdr
-        if ext_hdr is not None:
-            assert 'foo' in hdr
-
-    def teardown_class(self):
-        shutil.rmtree(self.outdir)
-
-
-@pytest.mark.skipif('not HAS_SCIPY')
 @pytest.mark.skipif('not HAS_SPECUTILS')
-class TestSpecutilsBridge:
+class TestSpecutilsBridgeSource:
     def test_from_spectrum1d_Empirical1D_source(self):
         import specutils
 
@@ -1266,70 +684,3 @@ class TestSpecutilsBridge:
         assert_quantity_allclose(
             integrated_flux, 1 * total_flux.unit, rtol=0.002)
         assert_quantity_allclose(integrated_flux, line_flux(spec), rtol=1e-5)
-
-    def test_from_spectrum1d_Empirical1D_bandpass(self):
-        import specutils
-
-        lamb = [1000, 5000, 10000] * u.AA
-        thru = [0, 1, -1] * units.THROUGHPUT
-        spec = specutils.Spectrum1D(spectral_axis=lamb, flux=thru)
-        with pytest.warns(AstropyUserWarning,
-                          match=r'contained negative flux or throughput'):
-            bp = SpectralElement.from_spectrum1d(spec, keep_neg=False)
-
-        w = bp.waveset
-        assert isinstance(bp.model, Empirical1D)
-        assert_quantity_allclose(w, lamb)
-        assert_quantity_allclose(bp(w), [0, 1, 0])
-
-    def test_from_spectrum1d_Empirical1D_bandpass_masked(self):
-        import specutils
-
-        lamb = [1000, 5000, 10000] * u.AA
-        thru = [0, 1, -1] * units.THROUGHPUT
-        mask = np.array([False, False, True])
-        spec = specutils.Spectrum1D(spectral_axis=lamb, flux=thru, mask=mask)
-        bp = SpectralElement.from_spectrum1d(spec, keep_neg=False)
-
-        w = bp.waveset
-        assert isinstance(bp.model, Empirical1D)
-        assert_quantity_allclose(w, [1000, 5000] * u.AA)
-        assert_quantity_allclose(bp(w), [0, 1])
-
-    def test_to_spectrum1d_Empirical1D_bandpass(self):
-        lamb = [1000, 5000, 10000] * u.AA
-        thru = [0, 1, 0]
-        bp = SpectralElement(Empirical1D, points=lamb, lookup_table=thru)
-        spec = bp.to_spectrum1d()
-
-        assert_quantity_allclose(spec.spectral_axis, lamb)
-        assert_quantity_allclose(spec.flux, thru)
-
-    def test_to_spectrum1d_Const1D(self):
-        thru = 0.88
-        bp = SpectralElement(Const1D, amplitude=thru)
-
-        with pytest.raises(exceptions.SynphotError) as e:
-            spec = bp.to_spectrum1d()
-        assert 'Provide wavelengths for sampling' in str(e.value)
-
-        w = [100, 500, 1000] * u.nm
-        spec = bp.to_spectrum1d(wavelengths=w)
-
-        assert_quantity_allclose(spec.spectral_axis, w)
-        assert_quantity_allclose(spec.flux, thru)
-
-    def test_to_spectrum1d_compound_bandpass(self):
-        from specutils.analysis import line_flux
-
-        box = SpectralElement(
-            Box1D, amplitude=0.5, x_0=5000 * u.AA, width=1 * u.AA)
-        bp = box * box
-        spec = bp.to_spectrum1d()
-
-        w = bp.waveset
-        integrated_thru = bp.integrate()
-        assert_quantity_allclose(spec.spectral_axis, w)
-        assert_quantity_allclose(spec.flux, bp(w))
-        assert_quantity_allclose(integrated_thru, 0.25 * u.AA)
-        assert_quantity_allclose(integrated_thru, line_flux(spec))
